Remove anonymous classes for snapshotters :{
