+++ src/Snapshot/CollectionSnapshot.php
@@ -11,6 +11,8 @@
 
 namespace Totem\Snapshot;
 
+use Totem\Snapshot;
+
 /**
  * Represents a snapshot of a collection
  *
@@ -19,8 +21,33 @@ namespace Totem\Snapshot;
  * @internal
  * @author Baptiste Clavié <clavie.b@gmail.com>
  */
-interface CollectionSnapshot
+final class CollectionSnapshot extends Snapshot
 {
+    /**
+     * Data mapper between the primary key and the real integer key of
+     * this collection
+     *
+     * @var int[]
+     */
+    private $link;
+
+    public function __construct($raw, array $data, array $link)
+    {
+        if (!is_array($raw) && !$raw instanceof Traversable) {
+            throw new InvalidArgumentException(sprintf('Expected a traversable, got %s', gettype($raw)));
+        }
+
+        parent::__construct($raw, $data);
+
+        $this->link = $link;
+    }
+
+    /** {@inheritDoc} */
+    public function isComparable(Snapshot $snapshot): bool
+    {
+        return $snapshot instanceof CollectionSnapshot;
+    }
+
     /**
      * Returns the original key for the primary key $primary
      *
@@ -28,6 +55,13 @@ interface CollectionSnapshot
      *
      * @throws InvalidArgumentException primary key not found
      */
-    public function getOriginalKey(string $primary): int;
+    public function getOriginalKey(string $primary): int
+    {
+        if (!isset($this->link[$primary])) {
+            throw new InvalidArgumentException(sprintf('The primary key "%s" is not in the computed dataset', $primary));
+        }
+
+        return $this->link[$primary];
+    }
 }
