+++ src/Snapshotter/CollectionSnapshotter.php
@@ -84,42 +84,7 @@ final class CollectionSnapshotter implements Snapshotter
             $link[$primary] = $key;
         }
 
-        return new class ($raw, $data, $link) extends Snapshot implements CollectionSnapshot {
-            /**
-             * Data mapper between the primary key and the real integer key of
-             * this collection
-             *
-             * @var int[]
-             */
-            private $link;
-
-            public function __construct($raw, array $data, array $link)
-            {
-                if (!is_array($raw) && !$raw instanceof Traversable) {
-                    throw new InvalidArgumentException(sprintf('Expected a traversable, got %s', gettype($raw)));
-                }
-
-                parent::__construct($raw, $data);
-
-                $this->link = $link;
-            }
-
-            /** {@inheritDoc} */
-            public function isComparable(Snapshot $snapshot): bool
-            {
-                return $snapshot instanceof CollectionSnapshot;
-            }
-
-            /** {@inheritDoc} */
-            public function getOriginalKey(string $primary): int
-            {
-                if (!isset($this->link[$primary])) {
-                    throw new InvalidArgumentException(sprintf('The primary key "%s" is not in the computed dataset', $primary));
-                }
-
-                return $this->link[$primary];
-            }
-        };
+        return new CollectionSnapshot($raw, $data, $link);
     }
 
     /**
